Chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "tavily-python>=0.7.23",
    "typing>=3.10.0.0",
    "uvicorn>=0.42.0",
]
